+++ src/models/Company.ts
